lights: Add ability for dual leds and correct battery state
Show correct battery state:
Android displays the battery as full even though it's only at 90%.
With this patch a posix thread will be started that'll make sure
the LED is amber even if android sets it to green if the battery
is not full yet.

Add ability for dual leds:
The Leo has only native support for shortly flashing the green
and amber led. With this patch the device will always have the LED
on and shortly switch to the other LED color if there is a
notification while the device is plugged in.